// sourcecode/BTrak.Packages/snova-canteen/module/module-components/src/lib/canteen/models/canteen-credit-model.ts
export class CanteenCreditModel {
    userIds: string[];
    amount: number;
    currencyId: string;
    canteenCreditId: string;
    userId: string;
    userName: string;
    userProfileImage: string;
    createdByUserId: string;
    currencyName: string;
    createdByUserName: string;
    creditedDate: Date;
    totalCount: number;
    createdByProfileImage: string;
}